// web/src/app/openapi/model/application.ts
/**
 * PIN
 * Personal Intelligent Network
 *
 * The version of the OpenAPI document: 8.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Application { 
    _id?: string;
    _updated?: string;
    _created?: string;
    name: string;
    description?: string;
}